requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
PyMySQL==1.1.1
python-docx==1.1.2
openai==1.40.0
httpx==0.27.0
pydantic==2.7.4
